BM-825: Update path filtering on CI to speed up PRs with changes to docs or infra (#563)
When making changes to infra or docs in particular, there is no need to
build and test the core Rust and Solidity code base. In promotion of
making more frequent, smaller PRs, that clear CI faster, this PR updates
the path filters to allow those PRs to merge without the main codebase
test suite.