BM-840: fix logic with estimation log and warn if broker behind (#580)
There was an underflow in the log if the estimation is before the
current time. This seems like a reasonable place to warn, as it
indicates that the broker may overlock orders. Possibly could be noisy
if tuned tightly, but airing on the side of caution. Happy to only warn
if above a margin of error also

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>
Co-authored-by: Victor Graf <victor@risczero.com>