BM-831: Order Generators auto deposit to market (#568)
When submitting requests off-chain, you must have balance pre-deposited
on the market. This means when we top up the order generators, we have
to manually send a deposit tx for each one, which we sometimes forget to
do. This change adds a convenience flag to auto-deposit when submitting
requests off-chain

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>