BM-637: Add error codes and alerts to the prover (#590)
BM-637

The goal is to be able to have custom alerts with different thresholds
for different types of errors.

Approach I took is to:
* Require our errors for each service implement trait `CodedError`,
requiring them to return an error code.
* Change each service to return their specific error type, rather than
just `anyhow::Error`. This ensures that errors returned bubbled up by
business logic contain error codes
* Modify the error messages to include the error code at the start of
the string.
* On the infra side, match on error codes to create metrics in
Cloudwatch.

I've included an initial set of alarms and thresholds, but these are
mostly just something to get started with, and will need to be modified
over time.