BM-861: Log error codes when printing errors using Debug trait. Add missing market monitor alarms (#598)
When logging an error using trait Debug's formatting using `{:?}` error
codes were not being logged, since the Debug derive is not aware of
them.

This PR adds a custom Debug macro that includes error codes. 

It also changes the logging behavior so that backtraces are not printed
unless the `RUST_BACKTRACE` environment variable is set to true. Imo the
backtraces are noisy and not particularly useful for debugging.

It also adds a custom alarm for event polling RPC errors, and a missing
alarm on unexpected errors in the market monitor