BM-874: Remove lock failed alarms, lower hyper_util log level (#610)
Lock Failed Alarms: Both these are expected when someone else beats us
to locking the request. No need to alert on them.

re: hyper_util, I moved the slasher logging to debug to debug the SQL
issue. On debug mode hyper_util logs "error" when a connection is
closed, which triggers our alarms, so moving it back to info.