BM-891: multiply anvil gas estimates in testing contexts (#619)
Not quite ideal to have different gas estimation logic for testing, but
currently unclear if based around anvil or if there is some edge case
that can be hit that changes the gas usage. This is a stop-gap to avoid
random failures in tests, as this has not been seen outside of tests
yet. 👀

Also changes the invocations of `BrokerBuilder::new_test` to keep around
the config file handle so it isn't dropped and causes error logs within
the tests.