BM-915: Adjust slasher + prover alarms. Install cargo for indexer pipeline (#634)
Assorted Ops Fixes:
* Change slasher error log alarm to alarm on "ERROR". Currently WARN
logs with the word error are triggering the alarm. Since we have retries
that when exhausted log with ERROR level, this change should not miss
issues
* Remove alarm on failing to get user provided image from ipfs. Add
metrics instead. We shouldn't alert on this, but could be useful to have
a metric on it.
* Indexer pipeline CodeBuild currently fails due to not having cargo
when running cargo-lambda. Installing cargo in the code builder