BM-921: Improve indexer (#640)
The indexer was calling too many times get_block_by_number, to get its
timestamp.
This PR adds a blocks table to store the block timestamp, whenever there
is a cache miss and attempt to retrieve it from the db whenever the
timestamp for a block is required.
Moreover refactor the starting block logic so that if the starting block
is set, use the maximum of the starting block and the last processed
block; otherwise, use the last processed block if it is greater than 0,
or the current block if the last processed block is 0. This is to ensure
that we don't process blocks that have already been processed and to
avoid processing blocks that are in the future.


It also tune down the log level