BM-926: Add fatal log on process exit, and alarms. Fix order generator topup (#644)
Noticed that the off-chain order generator has been crashing but not
alarming us. The error alarm there was misconfigured, so fixed that. But
also noticed we don't have alarms for when ecs tasks error out and exit.
Added those to all services.

Reason for off-chain order generator crashing is that it creates orders
with a high max price (e.g. > 10 ETH) and didn't have enough balance on
the market to submit the request. The auto deposit feature was set too
low to cover this, so increased it to 20 ETH.

I haven't tested this as this would require redeploying every service,
but fairly sure it should be ok ..