BM-931: Redeploy ECS tasks when secrets updated (#649)
When we change secrets a new ECS deployment is not triggered. This is
because changing the secret updates the value stored in Secret Manager,
but ECS only sees the Secret Arn, not the secret itself, so it doesn't
know it needs to redeploy.

To fix we pass a hash of the secrets as an env variable. Changes in env
variables _do_ trigger a redeployment.