BM-933: Fix missing locked event on the Indexer (#650)
This PR is a sort of hack for fixing the Indexer crashing after failing
to process a slashed event.
In case the DB has gaps, it might happen that when processing a slash
event the corresponding locked event is missing.
In that case we just return the Address::ZERO as prover address