BM-916: Deploy bento broker on aws (#636)
Deploys a Bento Broker to EC2. Uses instance local storage rather than
EFS/EBS. For updates, uses SSM to run a script that stops the broker,
pulls from Git, then restarts the service.

Few changes that were required for this:
* `justfile` now supports starting broker in attached mode rather than
detached. This allows `systemd` to monitor the broker service on EC2
* The setup script supports running in non-interactrive mode. This
allows the script to be run as part of the EC2 userdata startup process