BM-950: Tx timeouts in submitter as known errors (#663)
The transactions in the submitter can occasionally fail due to not being
confirmed in time. Given those transactions the same treatment as we
have for sending lock transactions, which is currently only alert us on
3 errors in an hour.

Also changing the default retries for batch submission down to 2.
Generally if a single retry here doesn't work a 3rd won't work either,
and retrying too much wastes time + gas