BM-958: add just installer to prover setup script (#671)
I noticed when running a prover on a fresh machine, that we recommend
you run the setup script, and then we ask to run a lot of different
`just` commands. If you're running this on a fresh machine, the script
handily installs Rust, and Docker and Nvidia etc. dependencies for you,
but not the `just` command runner.

It felt like it was missing, so I added it to the `setup.sh` script and
verified it worked on an Ubuntu 22.04 install.

---------

Co-authored-by: Austin Abell <austinabell8@gmail.com>