BM-769: Improve request building with an alloy-inspired RequestBuilder system (#515)
Inspired by `alloy`'s `TxFiller`, this PR introduced a `RequestBuilder`
trait along with an implementation built from a series of `Layer`
implementations. This `RequestBuilder` is integrated with the `Client`
and used to simplify the minimal implementation for new applications.

Note that the docs have been updated such that the snippets with build
and show some of the new patterns, but @sashaaldrick is working on a
more complete reframing around the new pattern.