BM-998: Add additional retries to Bonsai, adjust monitor alarms (#700)
* Adds a retry to preflighting only if preflighting fails with a 500
internal error (which should indicate the error was transient and it is
retryable). Note we already had retries set up for the call to create
the session, but the result of the session was not considered for
retrying
* During `find_and_monitor_proofs`, sometimes it runs before Redis has
finished bootstrapping. Adds a sleep and retry as a bit of a hack to
address this.
* Fixes an off-by-one bug in futures retry which meant that we would
retry less than configured
* Adds functionality to limit the max number of cycles generated by the
order generators. The variable numbers of cycles means that some proofs
take much longer than others to prove, which makes have a steady state
of fulfillments for our success rate metrics difficult.
* Building off this, I changed the off-chain OG to limit its cycles to a
small value so we have a steady stream of fulfillments. I left the
onchain OG as is.
* Various other fixes/tweaks to the alerts/thresholds