BM-1206: Avoid creating new proving sessions on retry (#808)
Currently the logic is handled such that if a proof is being generated
within the retry, it will create a new stark session on the next retry.
This can cause cascading failures due to not matching the estimation
logic.

Ideally in future it would be nice to estimate if we can create a new
session and try if possible, but would require a larger refactor.