BM-1213: fix dead links (hopefully forever) (#816)
fixing dead link on /developers/quick-start page that @nahoc pointed
out, I looked into adding a link checker. Quickly, I realized there is
already a link checker in `documentation/scripts/check-links.ts`, but
clearly wasn't working as intended.

Looking in there, it only checks `.md` files and not `.mdx` files which
is what we use so only the README.md was being checked.

This PR addresses this, and fixes the script; and fixes all the dead
links the scripts finds so that the CI passes :)