BM-1225: Make RPC errors in monitor/picker known. Raise threshold for submitter alarms (#825)
We've seen some RPC errors when getting wallet balances that are
currently Unexpected errors. Catching them and setting a higher
threshold.

Also raising the threshold on the market error alert in the submitter.
We retry in the submitter and often see recovery from those errors, so
no need to alert unless it fails multiple times.