BM-1244: Cache order IDs that have been priced to avoid duplicates (#842)
This is originally motivated by issues with a RPC provider sending
duplicates of on-chain events, but this can be useful otherwise to avoid
an order being spammed and re-priced multiple times. Also added a check
if locked before pricing, to avoid cases where there is a delay in
observing where an order is already locked.

~~Draft atm to rethink if there is a cleaner way to do this.~~ Edit:
I've opened for review as it's a bit opinionated, so feel free to
suggest an alternative!