BM-1310: Log session id for preflights, add function for getting proof prover (#894)
Seeing some strange behavior with the preflights, and would be much
easier to debug if we logged the session ids exactly when we create them
+ log them with the order id in the string so we can pull them with 1
log query.

The API change to the prover here doesn't make the most sense as we
shouldn't really pass in order id, but its the quickest way to get the
session ids. Happy to revert later for a cleaner solution.

Also adding a function for getting the prover of a fulfillment, which
will be useful for the proof indexer