BM-675: Update license headers to BSL and add copy of Apache license to repo root (#432)
This updates the header on files under the BSl license to point to the
appropriate license. This also adds a copy of the Apache license to the
repo, since many of the files are under that license. This also changes
the header of documentation, examples, and the slasher to indicate
Apache license.

---------

Co-authored-by: pote.eth <81638931+willpote@users.noreply.github.com>