BM-1356: Wipe bonsai DB, increase threshold for topping up stake (#921)
Few fixes for staging:

* The bonsai brokers are both erroring a lot with DB locked. Since its
been a while since we monitored the staging brokers so its unclear what
errors have built up in the DB and its general state, and since we want
to move off EFS anyway to solve this issue once and for all, I propose
we wipe the DB as a quick fix to get things rolling.

* One of the brokers is erroring due to low stake. This is happening due
to the threshold for topping it up being too low in the distributor.