BM-1363: Fix Ethereum Sepolia monitor success rate calculation (#929)
This adjusts the lambda to check the timestamp of the fulfillment rather
than the timestamp of the submission so that it gets counted within the
window of what is checked. This part was an autogenerated change, but
seems correct from looking it over.

I also increased the alert window from 30 mins to 60 as on graceful
stop, re-compilation, and then a lock and fulfill, that is very tight.
Hopefully reduces false positives.