BM-1380: Handle slash reverts and cases where someone else slashed (#935)
When we send a slash tx we did not check if the tx reverted, which meant
we would expect a ProverSlashed log event, then error out if its not
seen. In cases where someone else slashed the request, this would cause
issues.

Also some clippy failures and log level adjustments.