BM-1392: Reduce burn percentage to 20% (#944)
Currently 75% of stake gets burned, which feels excessive, especially as
the stake token is USDC. Lowering the burn percentage will provide more
incentive for secondary fulfillment, while still providing protection
against various DOS attacks.

There will need to be a second PR that we release after upgrade that
updates the broker to be aware of the new lower burn amount.