BM-1420: Unexpected error alarm threshold -> 2, lower traffic on staging, enable alarms on Base testnet (#966)
Various changes to make alerts less noisy, and some drive by fixes to
the slasher and logs:

1/ Big change is adjusting the unexpected error thresholds from 1 -> 2.
Eliminating all unexpected errors is challenging given how fast we move,
and the complexity of the system. Generally a single unexpected error is
OK and not a large cause for concern.

2/ Enable the high level alerts for Base Sepolia staging and prod.
Somehow these were missed.

3/ Reduce the amount of very large orders we send through staging, and
the threshold for triggering alarms there. Given it is only our Bento
prover (which we set max mcycle to 250 for) and Bonsai, and given all
our Bonsai provers are competing for resources, there can be cases under
load where the large orders don't on staging don't get picked up.

4/ Drive by fixes for the slasher. Theres a race condition possible
where we checked if the request is slashed, see it isn't, then someone
else slashes before we do. Adjusted the SDK to return a custom error
when this happens (the expected log doesn't get emitted)